Recent changes, including short-circuit caching

+++ application/libraries/pagecache.php
@@ -1,4 +1,4 @@
-<?php if ( ! defined('BASEPATH')) exit('No direct script access allowed'); 
+<?php
 
 /**
 * The entire reason for Chintzy to exist. This class should be loaded
@@ -15,6 +15,10 @@
 *    storage.
 */
 
+define('CACHE_HIT_NONE', 0);
+define('CACHE_HIT_DISK', 1);
+define('CACHE_HIT_MEMORY', 2);
+
 class Pagecache {
     // Default hash algorithm : MD5
     private $hash_algorithm = "md5";
@@ -29,17 +33,32 @@ class Pagecache {
     private $CI = NULL;
     
     // App path, because I'm lazy
-    private $APPPATH = APPPATH;
+    private $APPPATH = "";
     
     // Whether the caching mechanism should be enabled for this page
     public $enabled = TRUE;
+    
+    // Whether the cache successfully 'hit' an item
+    private $cache_hit = CACHE_HIT_NONE;
+    
+    // Generation timestamp
+    private $timestamp = NULL;
 
-    public function __construct() {
-        $this->CI =& get_instance();
+    public function __construct($app_path = '') {
+        if (function_exists("get_instance")) {
+            $this->CI =& get_instance();
         
-        $this->CI->config->load("cache");
-        $this->hash_algorithm = $this->CI->config->item("hash_algorithm", "cache");
-        $this->max_size = $this->CI->config->item("max_size", "cache");
+            $this->CI->config->load("cache");
+            $this->hash_algorithm = $this->CI->config->item("hash_algorithm", "cache");            
+            $this->max_size = $this->CI->config->item("max_size", "cache");
+        }
+    
+        $this->APPPATH = defined('APPPATH') ? APPPATH : $app_path;
+        $this->timestamp = time();
+    }
+    
+    public function hit() {
+        return $this->cache_hit;
     }
     
     public function get($key) {
@@ -49,23 +68,37 @@ class Pagecache {
     
         // Check APC/CodeIgniter built in cache functions first.
         $hashed_key = hash($this->hash_algorithm, $key);
-        $cache_data = $this->CI->cache->get($hashed_key);
+        $cache_data = $this->CI ? $this->CI->cache->get($hashed_key) : apc_fetch($hashed_key);
         
         if ($cache_data) {
             // Serve out this content
-            $this->CI->output->set_output("$cache_data");
-            return TRUE;
+            $this->cache_hit = CACHE_HIT_MEMORY;
+            if ($this->CI) {
+                $this->CI->output->set_output("$cache_data");    
+                return TRUE;
+            }
+            
+            // Return first element in array if it's an APC hit
+            $cache_data = $cache_data[0];
+            return $cache_data;
         }
         
         // Now check the caching directory in the filesystem
         if (file_exists("{$this->APPPATH}cache/$hashed_key.html")) {
             $output = file_get_contents("{$this->APPPATH}cache/$hashed_key.html");
-            $this->CI->cache->save($hashed_key, $output, $this->cache_expiry);
-            $this->CI->output->set_output("$output");
-            return TRUE;
+            $this->cache_hit = CACHE_HIT_DISK;
+            if ($this->CI) {
+                $this->CI->cache->save($hashed_key, $output, $this->cache_expiry);
+                $this->CI->output->set_output("$output");
+                return TRUE;
+            }
+            
+            // If we're not running inside CI, return the output.
+            return $output;
         }
         
         // Return false since we don't actually have any cached data
+        $this->cache_hit = FALSE;
         return FALSE;
     }
     
@@ -80,26 +113,60 @@ class Pagecache {
         }
     
         $hashed_key = hash($this->hash_algorithm, $key);
-        $this->CI->cache->save($hashed_key, "$data<!-- APC -->", $this->cache_expiry);
-        file_put_contents("{$this->APPPATH}cache/$hashed_key.html", "$data<!-- filecache -->");
+        
+        // Cache data according to its tier. Higher tier data is faster and more easily returned.
+        // Lower tier data is on slower storage, and the base level is "not currently cached".
+        
+        // If the cache hit the disk, but nothing higher tier than that, store in memory.
+        if ($this->cache_hit <= CACHE_HIT_DISK) {
+            $this->CI->cache->save($hashed_key, "$data<!-- mem {$this->timestamp} -->", $this->cache_expiry);
+        }
+        
+        // If the cache didn't hit, store on disk.
+        if ($this->cache_hit <= CACHE_HIT_NONE) {
+            file_put_contents("{$this->APPPATH}cache/$hashed_key.html", "$data<!-- disk {$this->timestamp} -->");
+        }
+        
+        // The cache chain at the end of an output file will appear like this:
+        /*
+        * No end comment: file was served from database/PHP generation and will be cached on
+        *    the next round in both memory and disk tiers
+        *
+        * mem comment: file was served from memory cache tier
+        * disk comment: file was served from disk cache tier and will be cached on the 
+        *    next round in memory tier
+        * disk + mem comment: file was served from memory cache tier, but originally retrieved
+        *    from the disk cache. Timestamps will indicate as much.
+        */
+    
     }
     
-    public function clear($key = NULL) {
+    public function clear($key = NULL, $tier = NULL) {
         if ($key) {
             $hashed_key = hash($this->hash_algorithm, $key);
             // Clear specific cache elements
-            $this->CI->cache->delete($hashed_key);
-            unlink("{$this->APPPATH}cache/$hashed_key.html");
-        } else {
-            // Clear entire cache
+            if (!$tier or $tier == CACHE_HIT_MEMORY) {
+                $this->CI->cache->delete($hashed_key);
+            }
+            if (!$tier or $tier == CACHE_HIT_DISK) {
+                @unlink("{$this->APPPATH}cache/$hashed_key.html");
+            }
+            return;
+        }
+
+        // Clear entire cache
+        if (!$tier or $tier == CACHE_HIT_MEMORY) {
             $this->CI->cache->clean();
+        }
+        
+        if (!$tier or $tier == CACHE_HIT_DISK) {
             if ($handle = opendir("{$this->APPPATH}cache")) {
                 while (FALSE !== ($file = readdir($handle))) {
                     if ($file === "." or $file === ".." or substr($file, -5) !== ".html") {
                         // Do nothing and don't go up one level
                     } else {
                         // Delete cache file
-                        unlink("{$this->APPPATH}cache/$file");
+                        @unlink("{$this->APPPATH}cache/$file");
                     }
                 }
                 closedir($handle);
